// app/models/customer.model.ts
export class Customer {
    kupacId?: number = 0;
    sifra?: string = '';
    naziv?: string = '';
    adresa?: string = '';
    email?: string = '';
    telefon?: string = '';
    pdvbroj?: string = '';
    idbroj?: string = '';
    gradId?: number = 0;
    gradNaziv?: string= '';
    klijentId?: number = 0;
    brojDana?: number = 0;
    brojBankovnogRacuna?: string = '';
}